Moved query_tx_deltas into shared crate.
